Make the 'else' part explicit in parse trees
Currently, it is rather cumbersome to detect the 'else' part of an 'if-else' construct in the parse tree because the 'else' part is flattened. This PR makes the parse tree more explicit by fields for the if-else-parts. This makes the grammar more robust and in line with other grammars like Java.