Fix call expression syntax (#27)
This PR fixes the `call_expression` and `dot_qualified_expression`
(former `navigation_expression`) syntax to fix the parse error issues
and wrong parse trees. The name changes were made to make the grammar
closer to the actual parser that used to the compiler than the reference
grammar.

Parse error in the delegation syntax, e.g., `by` keyword:

```
return object : A by a {
}

```

The current parser produces a parse error for this case as it tries to
parse it as a method call `a {}` on `return object : A b` which, of
course, doesn't make any sense.
Another problematic example, that doesn't lead to parse error, but
produces the wrong parse tree is this:

```
f as java.util.BitSet

```

Currently, this expression is parsed as `(((f as java).util).BitSet)`
because the priority of `as` and `.` is not working as declared in the
grammar. This expression should be parsed as `f as (java.utilBitSet)`.

A number of test cases added to cover these scenarios.