Make expression a visible node and super type (#34)
This PR makes the `expression` node visible in the parse tree and also
makes a super_type, so that it can be matched in queries, but will not
be present in the parse tree actually. This will help writing queries
that want to match any expression.

All existing tests pass without any modification.